moved testBAMReadBases to BAMFileWriterTest
